[PATCH] minor cleanup in ext3 code
The following patch is a _very minor_ cleanup in the ext3 code for
ext3_prepare_write().  It simply removes the setting of "handle" from
the current transaction handle, because "handle" is actually set again
a couple of lines later, where it properly allocates a new transaction
handle for this write.  The code removed in this patch is probably left
over from some development version of ext3 where the transaction handle
was started before ext3_prepare_write was called.

The only reason I saw this was because I was trying to find where the
handle was allocated for an ext3 file write, and at first glance it
didn't appear to be allocated anywhere...

BKrev: 3cccafd8Owjye6ChAnzEFiU1A5G63g