[PATCH] Allow truncate to discard unmapped buffers
The buffer state "uptodate, dirty, unmapped" is legal.  It occurs when
a mapped page with attached buffers which is over a hole is dirtied.

So discard_buffer() needs to be able to discard those buffers as well;
otherwise the page ends up clean, with dirty buffers and is unfreeable.

BKrev: 3cd5ad9dAlE-AnarenYtC7ZI4f8mhA