[PATCH] remove global_bufferlist_lock
This patch addresses Andrew's FIXME in buffer.c and adds a spinlock that
can be passed to the buffer list manipulation functions to the reiserfs
journal.  This obsoletes the cheks for lock beeing NULL in buffer.c and
the global_bufferlist_lock.

In addition osync_inode_list is changed to use list_for_each_prev as in 2.4
instead of a hand-crafted loop.

A little comment to the reiserfs folks:  your code would be _sooo_ much
easier to understand if you used temporary local variables for often
referenced fields..

BKrev: 3cd5c697jO32jdGJeABvbmDqcKNfBw