[PATCH] prefetching too far in mem copies
This patch from 2.4 makes sure we don't prefetch past the
end of a range to be copied (in case its at the end of a memrange)
i386 case looks safe already, we just weren't optimal for the last
chunk to be copied.

Andi. same change needed for x86-64.

    Dave.

BKrev: 3cd7fe8cwwze0kDI47amnmG08cKVIw