[PATCH] PNPBIOS / floppy conflict.
The PNPBIOS driver and the floppy driver both fight over reservation
of the floppy port. This patch makes things friendly again.
As the comment suggests, this fix could be done better, but until
someone steps forward to fix this, this is the best we have.

BKrev: 3cd7fec7-Ze5Z5pzBGQZ7RlrNeXSuQ