[PATCH] 2.4.19-pre8 uhci.c incorrect bit operations
This fixes up a couple of problems I came across while working on
uhci-hcd. There are a couple of places where shifts are used where they
shouldn't be and others where should be.

This cleans up a couple of cases and tidys it up.

The patch is relative to 2.4.19-pre8 and my other patches, but it's
alright to wait for 2.4.20. It should also be applied to 2.5.

BKrev: 3ce3ff7ax3xEdR6hf2QZQTg02u5JDg