[PATCH] Increase snd buffer size for UDP
I under-allocated the space that RPC needs for transmit on a UDP
socket.   We need lots of space to be allowed (though we may not use
much) so that a lack of un-commited transmit space will not cause new
requests to be blocked.

BKrev: 3ce6a560Ksl89_CpdnhUqEJ9k6f0dg