[PATCH] Change MD Superblock IO to go straight to submit_bio
The current code hits the page cache for the block device
which requires memory allocation which can sometimes cause
a deadlock (if it blocks the raid5d thread).

This code takes the page that holds the superblock, and
passes it to submit_bh in a suitable bio wrapper.

BKrev: 3ce6a56c0U9LriZDdCz8anUQqBtGBA