Merge http://linux-isdn.bkbits.net/linux-2.5.make into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/05/20 14:22:57-05:00 kai
kbuild: Speed up vmlinux build

The recursive build used to do the following: On entering a directory,
we would first call make in the subdirectories listed in $(subdir-y),
and then call make in the current directory again, with the target
'all_targets'.

The second invocation was used to make sure that the subdir makes completed
before trying to build the O_TARGET/L_TARGET in the current dir, since
this would link in objects in these subdirectories.

However, using correct dependencies achieves the same, i.e. just
make the objects in $(obj-y) which are not in the local subdirectory
depend on 'sub_dirs', which is the rule to descend into subdirs.

This patch actually halves the time "make vmlinux" takes when there's
nothing to do (which makes sense, as we save half of the make invocations)

2002/05/20 14:16:00-05:00 kai
kbuild: Remove generated .<object>.cmd files on 'make clean'

2002/05/20 14:14:59-05:00 kai
kbuild: Standardize building of init/*

There's no good reason to build the objects in init/ explicitly from
the top-level Makefile. Instead, handle init/ like every other
subdir, which also provides the automatic checks for a changed
command line etc.

BKrev: 3ce95302KhKAaZj9Nwy5k9kxv9JtRw