[PATCH] buffermem_pages removal (1/5)
With all work done by akpm in 2.5 Linus no more has a buffer cache
in the traditional sense.  Still we try to keep estimates about
what would be the buffer cache size by keeping the number of pages
indexed by block device inodes.  This is broken not only because the
old buffercache was also used for file data which is nowdays not
hashed to block device inodes and thus makes every user of this data
assume wrong numbers.  Second is is possible to use block device
pages not through the buffer_head interface (i.e. userspace
block device nodes, possibly JFS also soon).  In addition the atomic_t
used for this bookkepping (buffermem_pages) causes cacheline bouncing
on larger machines.

This is the first patch of a series to get rid of it.  It removes the
useless output of supposedly buffer pages in show_mem(), which is used
by the magic sysrq key code.

BKrev: 3cea92b3i3BzMLs6rWn1Hn7sbUr_sw