[PATCH] buffermem_pages removal (4/5)
This is the most discussion-worthy patch of the series:  Change the
meaning of si_meminfo->bufferram from "all pages in pagecache
backed by block devices" to "all pages in pagecache".

In the header/manpage is is documented as "Memory used by buffers",
but as the buffercache is gone I think the new meaning fits the
intention from pre-pagecache days much better.

BKrev: 3cea92catvH4C8dj2Sfc9H3rOrf-dw