Merge bk://linuxusb.bkbits.net/linus-2.5 into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/05/21 14:13:26-07:00 Andries.Brouwer
[PATCH] usb-storage

Created the first of a series of usb-storage patches
this afternoon. Please find it below or on ftp.XX.kernel.org
under people/aeb/2.5.16-us-patch .
It removes 5% of the usb-storage code, but should, if I made
no mistake, not affect the behaviour of the code.

(This is the introduction of raw_bulk.c to hold USB-bulk
and scatter-gather code that was repeated three-five times
in various drivers. The next patch is smartmedia.c.)

2002/05/21 14:13:05-07:00 greg
[PATCH] USB Makefile bug fix

fixed problem with UHCI_HCD drivers set to 'y' were not causing the build
to go into the drivers/usb/host directory.

2002/05/21 14:12:49-07:00 petkan
[PATCH] USB pegasus driver, new vendor and device id.

new vendor and device ID

BKrev: 3ceacedeHr7mhu3-PiGNDz4NCbeJpQ