Merge home.transmeta.com:/home/torvalds/v2.5/preemption into home.transmeta.com:/home/torvalds/v2.5/linux
2002/05/21 18:00:36-07:00 rml
[PATCH] remove preempt_disable from pdflush

It seems by the comments the statements were only there to prevent the
thread from finding itself as TASK_RUNNING after a kernel preemption.
Since we reverted to the original preemption behavior (which does not do
that) it should be safe.  Looking over the code confirms - it seems to
not be doing anything evil and is properly locked otherwise.

BKrev: 3ceaedc5AI7PXt_zKV1J6rRoUzo1HA