[PATCH] direct-to-BIO writeback
Multipage BIO writeout from the pagecache.

It's pretty much the same as multipage reads.  It falls back to buffers
if things got complex.

The write case is a little more complex because it handles pages which
have buffers and pages which do not.  If the page didn't have buffers
this code does not add them.

BKrev: 3cf2e7a24mMdOeknHlBKjmvd3lpi6g