[PATCH] avoid sys_sync livelocks
This makes sure that sys_sync() will terminate.  It counts up the
number of dirty pages in the machine and will refuse to write out more
than 1.25 times this number of pages.  This function is called twice
on the sys_sync() path, so the kernel will actually write 2.5x the number
of initially-dirty pages before giving up.

BKrev: 3cf2e827ZegE5PdGOcXoQyfrOMe8Mw