Merge signal arch cleanups
2002/05/28 19:52:44-07:00 sfr
[PATCH] consolidate errno definitions

Just remove duplicates among the asm-*/errno.h.

2002/05/28 19:52:39-07:00 sfr
[PATCH] consolidate do_signal

11 out of our 17 architectures have basically the same code
in arch/../kernel/signal.c:do_signal.  This patch creates a
common function for that bit of code and uses it in the places
it can be.

The 2.5.15 version of this patch builds and runs on i386 and PPC and has
been briefly looked at by the CRIS, PARISC, PPC64 and x86_64 maintainers.

As a bonus, this fixes the "ignore SIGURG" bug for 9 more architectures
(i386 and PPC already were fixed).

2002/05/28 19:52:34-07:00 sfr
[PATCH] consolidate generic peices of the siginfo structures and associated stuff

This patch creates asm-generic/siginfo.h and uses it to remove a
lot of duplicate code in the various asm-*/siginfo.h files.  Some
if it is a little ugly, but I think it will be worth it just to
help us eliminate some of the bugs that have come from code copying.

BKrev: 3cf443e4J4kZmnH04KbTWNvvx07Fxg