Merge bk://bk.arm.linux.org.uk into home.transmeta.com:/home/torvalds/v2.5/linux
2002/05/30 00:16:58+01:00 rmk
[ARM] Remove hard coded per-architecture memory, ramdisk and initrd
We have a fair number of command line options which can be used
to pass information about the target to the kernel.  We also provide
a default command line which the kernel will use in absence of one
provided by the boot loader.  We pass the hard coded options via the
command line now.

2002/05/29 23:55:34+01:00 rmk
[ARM] Manual merge.

2002/05/29 23:38:24+01:00 rmk
[ARM] Convert for() delay loops to udelay()

2002/05/29 16:21:22+01:00 rmk
[ARM] Remove an extraneous load from atomic ops
atomic.h was generating some extra loads that aren't required.  In
addition, it was needlessly performing various tests inside the
atomic region.  This change fixes both these issues.

2002/05/29 16:06:54+01:00 rmk
[ARM] Context switch improvements
 - We don't need to save the CPSR.
 - Rearrange thread_info members so we can pull the fields out of
   thread_info more efficiently.
 - Allocate a couple of extra words for CPU specific context saving
   (eg, for Xscale ACC registers)
 - Always leave 8 bytes free at the top of the kernel stack.  This
   prevents the stack becoming completely empty when do_exit() is
   called from an exiting nfsd() thread, and causing the wrong
   pointer to be returned from current_thread_info()

2002/05/29 12:45:22+01:00 rmk
[ARM] cpufreq_init takes low and high frequency limits.

2002/05/29 12:34:31+01:00 rmk
[ARM] Fix makefiles for drivers/acorn/{block,char}/

2002/05/29 12:12:21+01:00 rmk
[ARM] Remove victor machine type
This machine type is no longer maintained and is falling into bitrot
heaven.

2002/05/26 01:04:52+01:00 rmk
[ARM] Tidy up context switch save area initialisation for fork
Fix a minor bug where we jump to ret_from_fork with IRQs enabled.

2002/05/26 00:51:36+01:00 rmk
[ARM] 2.5.18 Update ARM for TLB shootdown changes
We basically do the same as Sparc64.

2002/05/26 00:42:40+01:00 rmk
[ARM] Remove redundant instruction cache code from decompressor.
Now that we cleanly support handling of the cache for various ARM
processors in head.S, the per-processor class include file, and
associated inline function is no longer required.  This changeset
removes these files and the function call.

BKrev: 3cf5aba4xKHSIXu4V-RP8QeoJbfg1Q