[PATCH] trivial keyboard driver patch
This makes it possible to run the ia64 kernel both on platforms with the
legacy keyboard controller and those without (as is the case for the hp
zx1 platform, for example).  The new code is inside #ifdef CONFIG_IA64,
though it would probably be a good idea to enable it unconditionally.
The patch is by Alex Williamson.

BKrev: 3cf7bd37KQd8hNJ4tQ1cP1M-xPAP8Q