[PATCH] fix 2.5.20 ramdisk
2.5.20 seems to be incapable of executing binaries in a ramdisk-based
root filesystem.  The ramdisk in question is an ext2fs, with a 1K
block size loaded via the compressed ramdisk loader in do_mounts().

It appears that, in the case of a 1K block sized filesystem, we attempt
to read two 512-byte sectors into a BIO vector.  The first one is copied
into the first 512 bytes.  The second sector, however, is copied over
the first 512 bytes.  Obviously not what we really want.

Rev. 2, slightly cleaned up:

BKrev: 3cfceac0SFIvsUUxZHRxPINh0l0Yjg