[PATCH] fs/locks.c: Only yield once for flocks
This patch removes the annoying and confusing `wait' argument from
many places.  The only change in behaviour is that we now yield once
when unblocking other BSD-style flocks instead of once for each lock.

This slightly improves the semantics for userspace.  Before, when we
had two tasks waiting on a lock, the first one would receive the lock.
Now, the one with the highest priority receives the lock.

BKrev: 3d028ea4WkbBO5j4RHWuhKFJp1DZCA