Merge http://jfs.bkbits.net/linux-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/06/07 14:02:18-05:00 shaggy
JFS: remove obsolete declaration

2002/06/07 13:58:51-05:00 shaggy
JFS: Add hch's copyright.

He has added significant content to these files.

2002/06/07 13:54:51-05:00 shaggy
JFS: Fix structure alignment problem on 64-bit machines

The code in xtLog assumes that the structures xdlistlock_t and
maplock_t are the same size, as they are overlaid.  On 64-bit
machines, this was not the case.

Ensure that the xdlistlock_t is 16 bytes in length regardless of
the size of a pointer.

BKrev: 3d0292b5rLZ_fVyazf2kpJ3uyLQTEg