kbuild: Rearrange Rules.make
This patch doesn't change anything but the order (and some comments) in
Rules.make. The reason for that is that we use Rules.make for three
different purposes, fastdep, modules_install and the actual building.

As an example, we don't need all the dependencies and saved command lines
when installing modules, so it'll speed up this pass to only include the
rules which will actually get used.

BKrev: 3d03ba72s2ykgD9NBkR2dKVxGrFo8A