[PATCH] pci dma patch rediffed for 2.5.21
This is the same patch as was posted against 2.4.19-pre10: pci_map_page
was missing a cast on x86, which resulted in the high 32 bits of an
address being silently discarded.  This patch fixes that by casting
the page number before multiplying it out.

BKrev: 3d055f94lbM_1t7pXfrsGWmA12Rvaw