[PATCH] ns83820.c update to 0.18
This patch fixes a highmem issue with ns83820.c when used with the
pci dma fix for x86.  The patch was successfully tested on a highmem
machine with 5GB of ram and a file in high memory.  Cruft from the
pre-64 bit pci dma era is also removed.

BKrev: 3d05601ax-ZEixLhqP4w_iTdgOu1wQ